+++ src/main/java/com/weatherwhere/airservice/repository/airrealtime/StationNameRepository.java
@@ -1,6 +1,6 @@
-package com.weatherwhere.airservice.repository;
+package com.weatherwhere.airservice.repository.airrealtime;
 
-import com.weatherwhere.airservice.domain.RealTimeAirEntity;
+import com.weatherwhere.airservice.domain.airrealtime.RealTimeAirEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationNameRepository extends JpaRepository<RealTimeAirEntity, String> {
