+++ src/main/java/com/weatherwhere/airservice/service/airrealtime/RealTimeAirServiceImpl.java
@@ -1,10 +1,10 @@
-package com.weatherwhere.airservice.service;
+package com.weatherwhere.airservice.service.airrealtime;
 
 
-import com.weatherwhere.airservice.domain.RealTimeAirEntity;
-import com.weatherwhere.airservice.dto.RealTimeAirDto;
-import com.weatherwhere.airservice.repository.RealTimeAirRepository;
-import com.weatherwhere.airservice.repository.StationNameRepository;
+import com.weatherwhere.airservice.domain.airrealtime.RealTimeAirEntity;
+import com.weatherwhere.airservice.dto.airrealtime.RealTimeAirDto;
+import com.weatherwhere.airservice.repository.airrealtime.RealTimeAirRepository;
+import com.weatherwhere.airservice.repository.airrealtime.StationNameRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -19,6 +19,7 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.text.ParseException;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 
 @Service
@@ -61,8 +62,7 @@ public class RealTimeAirServiceImpl implements RealTimeService {
                 "&pageNo=1" +
                 "&numOfRows=1" +
                 "&returnType=json" +
-                "&serviceKey=" + "LKiVUvMq5P2zZ88FZoOq/h0k9y98k2pEdRcJSheoYPwZxYlcaGkQugApuMndBS0dqRg1QeziMPwW9rbVvRIcRA==" +
-                //System.getProperty("AIR_FORECAST_SERVICE_KEY_DE") +
+                "&serviceKey=" + System.getProperty("AIR_FORECAST_SERVICE_KEY_DE") +
                 "&ver=1.0";
         String jsonString = restTemplate.getForObject(apiUrl, String.class);
         try {
@@ -131,6 +131,11 @@ public class RealTimeAirServiceImpl implements RealTimeService {
         return "성공";
     }
 
+    @Override
+    @Transactional
+    public RealTimeAirEntity getDBData(String stationName){
+        return realTimeAirRepository.findById(stationName).orElseThrow(() -> new NoSuchElementException());
+    }
 
     //DB에 stationName을 csv에서 읽어와 저장하는 메서드
     private final StationNameRepository stationNameRepository;
